% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{summarize_daymet}
\alias{summarize_daymet}
\title{create Daymet summaries}
\usage{
summarize_daymet(x, var, days, years = 1980:2022, fun = mean)
}
\arguments{
\item{x}{a daymetr object}

\item{var}{character string giving desired weather variable}

\item{days}{numeric vector of desired days, or a function of a daymetr object
that returns a vector of days}

\item{years}{numeric vector of desired years}

\item{fun}{summary function taking vector input and returning scalar output}
}
\value{
a tibble giving the summary function of \code{var} over the
requested \code{days}, for each requested year in \code{years}, for each
site in \code{x}
}
\description{
create Daymet summaries
}
